--- source/features/rgh-welcome-issue.css
@@ -1,29 +1,0 @@
-/**
-
-!!!!!! Don't forget to add `[data-url]` selectors or else it will affect unrelated pages
-
-.gh-header-title is to exclude the sticky header
-
-*/
-
-[data-url^='/refined-github/refined-github/issues/3543/show_partial'] :is(
-.gh-header-meta, /* Issue details */
-.gh-header-actions, /* "New issue" button */
-.gh-header-title .color-fg-muted /* Issue number */
-),
-[data-url^='/refined-github/refined-github/issues/3543/partials'] :is(
-	.js-comment-edit-history, /* "Edited" badge */
-	relative-time /* Comment time */
-),
-[data-url^='/refined-github/refined-github/issues/3543/show_partial?partial=issues%2Fsidebar'] /* Sidebar */ {
-	display: none !important;
-}
-
-@media (width >= 768px) {
-	[data-url^='/refined-github/refined-github/issues/3543/show_partial']
-		.gh-header-title
-		.js-issue-title {
-		font-size: 5vw !important;
-		font-weight: 300;
-	}
-}
