+++ source/options.html
@@ -12,7 +12,7 @@
 	<details id="info" open>
 		<summary><img src="icon.png" alt="" width="24" height="24"> Refined GitHub</summary>
 		<p>
-			Visit the <a href="https://github.com/refined-github/refined-github/issues/3543">welcome page</a> to learn about updates, debugging, and GitHub Enterprise.
+			Visit the <a href="https://github.com/refined-github/refined-github/wiki">welcome page</a> to learn about updates, debugging, and GitHub Enterprise.
 			You can <a href="https://chrome.google.com/webstore/detail/refined-github/hlepfoohegkhhmjieoechaddaejaokhf/reviews" id="rate-link">rate Refined GitHub</a> to help others find it.
 			Follow or sponsor <a href="https://github.com/sponsors/fregante">@fregante</a> if Refined GitHub helps you work more efficiently. 🍻
 		</p>
