+++ source/refined-github.ts
@@ -176,7 +176,7 @@ import './features/no-unnecessary-split-diff-view.js';
 import './features/list-prs-for-branch.js';
 import './features/select-notifications.js';
 import './features/clean-repo-tabs.js';
-import './features/rgh-welcome-issue.js';
+import './features/rgh-options-link.js';
 import './features/same-branch-author-commits.js';
 import './features/prevent-pr-merge-panel-opening.js';
 import './features/rgh-improve-new-issue-form.js';
